Merged PR 28460: Add functional test miniport
As a part of the NL/FL testing strategy, we need a way forward for
devs to easily be able to add packet-level or OID/status-level test
cases to test automation. The established way in the past was a
solution involving .wsf scripts, fnShell, and the Sparta test
miniport. While it was initially preferred to re-use at least the
Sparta test miniport for the sake of minimizing test related
technical debt, the ickiness factor proved to be too much and a
backup option of porting a newer test miniport was chosen. This PR
introduces the newer test miniport, ported from xdp-for-windows.

https://microsoft.sharepoint.com/:w:/r/teams/STACKTeam-CoreNetworkingMobileConnectivityPeripheralsStackSe/_layouts/15/Doc.aspx?sourcedoc=%7BBD44D0B4-7165-4F84-9D8F-F5DF105C43C7%7D&file=NL_testing_strategy_Zn.docx&action=default&mobileredirect=true&share=IQG00ES9ZXGET52P9d8QXEPHARjggINmJJNUO-l2oQyr24Y

- Ported the XDP functional test miniport from xdp-for-windows
- Re-organized and deleted stuff to have it make more sense outside
  of XDP context
- Added some basic tests to verify at least some parts work